# api/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Imam Hadi Network API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/loans": {
      "get": {
        "tags": [
          "loans"
        ],
        "summary": "Paginated loan list with year/topic/status/person/liaison filters",
        "operationId": "list__api_loans_get",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 1500,
                  "minimum": 1300
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "topic_ids",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated topic ids, e.g. '3,7'",
              "title": "Topic Ids"
            },
            "description": "Comma-separated topic ids, e.g. '3,7'"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(active|settled)$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "active | settled",
              "title": "Status"
            },
            "description": "active | settled"
          },
          {
            "name": "borrower_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Borrower Id"
            }
          },
          {
            "name": "lender_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lender Id"
            }
          },
          {
            "name": "liaison",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Liaison"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Substring match on loan_number",
              "title": "Q"
            },
            "description": "Substring match on loan_number"
          },
          {
            "name": "due_within_days",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 3650,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only loans with an unpaid installment due within N days from today",
              "title": "Due Within Days"
            },
            "description": "Only loans with an unpaid installment due within N days from today"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(loan_number|year|total|remaining)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          },
          {
            "name": "sort_dir",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "default": "asc",
              "title": "Sort Dir"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_LoanListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/loans/{loan_id}": {
      "get": {
        "tags": [
          "loans"
        ],
        "summary": "Loan detail: borrowers, lenders, installments",
        "operationId": "detail_api_loans__loan_id__get",
        "parameters": [
          {
            "name": "loan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Loan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoanDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InstallmentRef": {
        "properties": {
          "due_persian_year": {
            "type": "integer",
            "title": "Due Persian Year"
          },
          "due_persian_month": {
            "type": "integer",
            "title": "Due Persian Month"
          },
          "due_day_of_month": {
            "type": "integer",
            "title": "Due Day Of Month"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "status": {
            "$ref": "#/components/schemas/InstallmentStatus"
          }
        },
        "type": "object",
        "required": [
          "due_persian_year",
          "due_persian_month",
          "due_day_of_month",
          "amount",
          "status"
        ],
        "title": "InstallmentRef"
      },
      "InstallmentStatus": {
        "type": "string",
        "enum": [
          "paid",
          "unpaid"
        ],
        "title": "InstallmentStatus",
        "description": "Repayment installment state.\n\nDerived from the cell fill colour in the source xlsm (``#00B050`` = paid,\nanything else = unpaid).  Phase 2 will add intermediate states once\nactual payment dates are recorded."
      },
...
      "LenderPartyRef": {
        "properties": {
          "party_id": {
            "type": "integer",
            "title": "Party Id"
          },
          "person": {
            "$ref": "#/components/schemas/PersonRef"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "paid": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Paid"
          },
          "remaining": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Remaining"
          },
          "installments": {
            "items": {
              "$ref": "#/components/schemas/InstallmentRef"
            },
            "type": "array",
            "title": "Installments"
          }
        },
        "type": "object",
        "required": [
          "party_id",
          "person",
          "amount",
          "paid",
          "remaining",
          "installments"
        ],
        "title": "LenderPartyRef"
      },
      "LoanDetailResponse": {
        "properties": {
          "loan": {
            "additionalProperties": true,
            "type": "object",
            "title": "Loan",
            "description": "Loan-level fields: id, persian_year, loan_number, channel_number, total_amount, liaison_label, description"
          },
          "topic": {
            "$ref": "#/components/schemas/TopicRef"
          },
          "guarantor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "borrowers": {
            "items": {
              "$ref": "#/components/schemas/LoanPartyRef"
            },
            "type": "array",
            "title": "Borrowers"
          },
          "lenders": {
            "items": {
              "$ref": "#/components/schemas/LenderPartyRef"
            },
            "type": "array",
            "title": "Lenders"
          },
          "totals": {
            "$ref": "#/components/schemas/LoanTotals"
          }
        },
        "type": "object",
        "required": [
          "topic",
          "guarantor",
          "borrowers",
          "lenders",
          "totals"
        ],
        "title": "LoanDetailResponse",
        "description": "Response shape for /api/loans/{id} (DESIGN.md §5.2)."
      },
      "LoanListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "persian_year": {
            "type": "integer",
            "title": "Persian Year"
          },
          "loan_number": {
            "type": "string",
            "title": "Loan Number"
          },
          "channel_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Number"
          },
          "topic_name": {
            "type": "string",
            "title": "Topic Name"
          },
          "borrower_name": {
            "type": "string",
            "title": "Borrower Name"
          },
          "liaison_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Liaison Label"
          },
          "total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total"
          },
          "paid": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Paid"
          },
          "remaining": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Remaining"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "settled"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "persian_year",
          "loan_number",
          "channel_number",
          "topic_name",
          "borrower_name",
          "liaison_label",
          "total",
          "paid",
          "remaining",
          "status"
        ],
        "title": "LoanListItem",
        "description": "Row on the loans list page (DESIGN.md §6.2.4)."
      },
      "LoanPartyRef": {
        "properties": {
          "party_id": {
            "type": "integer",
            "title": "Party Id"
          },
          "person": {
            "$ref": "#/components/schemas/PersonRef"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "party_id",
          "person",
          "amount"
        ],
        "title": "LoanPartyRef"
      },
      "LoanTotals": {
        "properties": {
          "total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total"
          },
          "paid": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Paid"
          },
          "remaining": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Remaining"
          },
          "settled": {
            "type": "boolean",
            "title": "Settled"
          }
        },
        "type": "object",
        "required": [
          "total",
          "paid",
          "remaining",
          "settled"
        ],
        "title": "LoanTotals"
      },
...
      "Page_LoanListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LoanListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[LoanListItem]"
      },
...
      "PersonRef": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "phone"
        ],
        "title": "PersonRef",
        "description": "Minimal nested person reference used in profile detail responses."
      },
...
      "TopicRef": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "TopicRef"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}